java commit

+++ JAVA/JavagroupA/src/book5/ch3/FactorialCompare.java
@@ -3,7 +3,7 @@ package book5.ch3;
 public class FactorialCompare {
 
     public static void main(String[] args) {
-      long n = 5;
+      int n = 5;
      long nonRecFactorial =  factorial_nRecursive(n);
      System.out.println("factorial of " + n + "is " + nonRecFactorial);
     }
